+++ src/Game/Shibboleth_HoldingLoader.cpp
@@ -25,8 +25,8 @@ THE SOFTWARE.
 #include <Shibboleth_IFileSystem.h>
 #include <Shibboleth_Utilities.h>
 #include <Shibboleth_IApp.h>
-#include <Gleam_IModel.h>
 #include <Gaff_ScopedExit.h>
+#include <Gleam_IModel.h>
 
 NS_SHIBBOLETH
 
@@ -38,7 +38,7 @@ HoldingLoader::~HoldingLoader(void)
 {
 }
 
-Gaff::IVirtualDestructor* HoldingLoader::load(const char* file_name, unsigned long long user_data, HashMap<AString, IFile*>& file_map)
+Gaff::IVirtualDestructor* HoldingLoader::load(const char* file_name, uint64_t user_data, HashMap<AString, IFile*>& file_map)
 {
 	AString fname(file_name);
 	assert(file_map.hasElementWithKey(fname));
