+++ src/Game/include/Shibboleth_RenderTargetLoader.h
@@ -32,7 +32,7 @@ public:
 	RenderTargetLoader(void);
 	~RenderTargetLoader(void);
 
-	Gaff::IVirtualDestructor* load(const char* file_name, unsigned long long, HashMap<AString, IFile*>& file_map);
+	Gaff::IVirtualDestructor* load(const char* file_name, uint64_t, HashMap<AString, IFile*>& file_map);
 
 	GAFF_NO_COPY(RenderTargetLoader);
 	GAFF_NO_MOVE(RenderTargetLoader);
